- implemented Symbolic Link - improved multiline edit - fixed bugs

+++ nsh.sh
@@ -195,10 +195,12 @@ nshcp() {
                 local cmp= && $(command diff --brief "$s" "$d" &>/dev/null) && cmp=', same file'
                 #echo -e "$NSH_PROMPT "$'\e[4m'"${d/$HOME\//$tilde\/}"$'\e[0m'" already exists ($(get_hsize $ssize) --> $(get_hsize $dsize)$cmp)\e[K"
                 echo -e "$NSH_PROMPT "$(print_filename "$d")" already exists ($(get_hsize $ssize) --> $(get_hsize $dsize)$cmp)\e[K"
-                local ans="$(menu --popup 'Overwrite' 'Overwrite all' 'Skip' 'Skip all' 'Rename' 'Cancel' --key o 'echo Overwrite' --key s 'echo Skip' --key r 'echo Rename' --footer "$(draw_shortcut o Overwrite s Skip r Rename)" --cursor $'\e[31;100m>' --sel-color '0;1;100')"
                 while true; do
+                    local ans="$(menu --popup 'Overwrite' 'Overwrite all' 'Skip' 'Skip all' 'Rename' 'Cancel' --key o 'echo Overwrite' --key s 'echo Skip' --key r 'echo Rename' --footer "$(draw_shortcut o Overwrite s Skip r Rename)" --cursor $'\e[31;100m>' --sel-color '0;1;100')"
                     case "$ans" in
-                        'Overwrite') ;;
+                        'Overwrite')
+                            echo "    $s --> $(print_filename "$d")"
+                            ;;
                         'Overwrite all') overwrite_all=yes;;
                         #'Skip') return;;
                         'Skip all') skip_all=yes && return;;
@@ -1064,13 +1066,13 @@ git_fix_conflicts() {
     [[ $# -gt 0 ]] && echo "$@"
     while true; do
         idx="$(for f in "${files[@]}"; do
-            [[ $(grep -c '<<< HEAD' "$f" 2>/dev/null) -gt 0 ]] && echo "@@$f" || echo "$f"
+            [[ $(grep -c '^<\+ HEAD' "$f" 2>/dev/null) -gt 0 ]] && echo "@@$f" || echo "$f"
         done | menu --popup --sel-color 4 --cursor $'\e[31;100m>' --sel-color '0;1;100' --return-idx --accent "@@" 31 32)"
         [[ -z "$idx" ]] && break
         $NSH_DEFAULT_EDITOR "${files[$idx]}"
         local cont=false
         for f in "${files[@]}"; do
-            [[ $(grep -c '<<< HEAD' "$f" 2>/dev/null) -gt 0 ]] && cont=true && break
+            [[ $(grep -c '^<\+ HEAD' "$f" 2>/dev/null) -gt 0 ]] && cont=true && break
         done
         if [[ $cont == false ]]; then
             echo -n 'All conflicts were fixed. Apply the changes to continue? (Y/n) ' && get_key KEY
@@ -1116,7 +1118,7 @@ git_log() {
                         fi
                     fi
                 fi
-            done < <(git log --follow "$@" 2>/dev/null)
+            done < <(eval "git log --follow "$@" 2>/dev/null")
         }
         commit="$(gather_commit "$@" | menu --header ' CommitID     Log' --footer "+ $(draw_shortcut TAB Preview ENTER Checkout \/ Search z Zoom)" --popup --sel-color 7 --preview git_commit_preview --searchable $mopt)"
         commit="${commit%% *}"
@@ -1338,12 +1340,13 @@ play2048() {
 # main function
 ############################################################################
 nsh() {
-    local version='0.1.20220613'
+    local version='0.1.20220921'
     local nsh_mode=
     local subprompt=
     local STRING=
     local STRING_SUGGEST=
     local PRESTRING=
+    local STRINGBUF=
     local selected=()
     local tilde='~'
     local dirs
@@ -1490,7 +1493,9 @@ nsh() {
         hide_cursor
 
         # ps
-        if [[ -z $subprompt ]]; then
+        if [[ -n $STRINGBUF ]]; then
+            subprompt="$NSH_PROMPT "
+        elif [[ -z $subprompt ]]; then
             local prefix=
             [[ -n $nsh_mode ]] && prefix=$'\e[37;45m'"[$nsh_mode]"
             if [[ -n "$NSH_PROMPT_PREFIX" ]]; then
@@ -2011,7 +2016,7 @@ nsh() {
             draw_help_line "Press $NSH_COLOR_SH1 v \e[0m to edit the file using the default editor. To check the default editor, see NSH_DEFAULT_EDITOR in config.\e[K"
             draw_help_line "See the table of important keyboard shortcuts below:\e[K"
             local c='Less' && [[ $show_all -eq 0 ]] && c='All'
-            draw_shortcut_ml "F2" "Rename" "F5" "Copy" "F6" "Move" "F7" "Mkdir" "F10" "Config" "g" "Git" "y" "Yank" "r" "Refresh" "i" "Rename" "I" "Touch" "dd" "Delete" "m" "Mark" "'" "Bookmarks" ';' "Commands" "Tab" "View" ":" "Shell" "/" "Search" "^G" "Grep" "." "Show$c" "s" "Sort" "~" "Home" "2" "2048"
+            draw_shortcut_ml "F2" "Rename" "F5" "Copy" "F6" "Move" "F7" "Mkdir" "F10" "Config" "g" "Git" "y" "Yank" "r" "Refresh" "i" "Rename" "I" "Touch" "dd" "Delete" "m" "Mark" "'" "Bookmarks" 'L' "Ln" ';' "Commands" "Tab" "View" ":" "Shell" "/" "Search" "^G" "Grep" "." "Show$c" "s" "Sort" "~" "Home" "2" "2048"
             show_cursor
             get_key
             disable_line_wrapping
@@ -2136,11 +2141,11 @@ nsh() {
         fi
     }
     open_file() {
-        if [ -d "$1" ]; then
-            command cd "$1"
+        local fname="${1/#$tilde\//$HOME/}"
+        if [ -d "$fname" ]; then
+            command cd "$fname"
             update
         else
-            local fname="$1"
             eval "[[ -e $fname ]] && echo" &>/dev/null || fname="\"$fname\""
             if [[ "$fname" == *.py ]]; then
                 subshell "python $fname "
@@ -2424,16 +2429,17 @@ nsh() {
         local yfocus
         local yy
         local ylast_item
-        local op='cp'
-        [[ $1 == 'mv' ]] && op='mv'
+        local op="$1"
         yopen() {
             path="$(command cd "$1/"; pwd)"
             [[ $path == / ]] && path=
             move_cursor "1;$((list_width+1))"
-            if [[ $op == 'mv' ]]; then
-                printf '\e[30;42m\e[K| Move to: %s\e[0m' "${path/#$HOME\//$tilde/}"
+            if [[ $op == cp ]]; then
+                printf '\e[30;42m\e[K| Copy to: %s\e[0m' "$(print_filename "$path")"
+            elif [[ $op == mv ]]; then
+                printf '\e[30;42m\e[K| Move to: %s\e[0m' "$(print_filename "$path")"
             else
-                printf '\e[30;42m\e[K| Copy to: %s\e[0m' "${path/#$HOME\//$tilde/}"
+                printf '\e[30;42m\e[K| Symbolic link\e[0m'
             fi
             dirs=()
             files=()
@@ -2471,6 +2477,7 @@ nsh() {
         ydraw() {
             local d=${#dirs[@]}
             local w=$((COLUMNS-list_width))
+            [[ $((yy+max_lines)) -lt $yfocus ]] && yy=$((yfocus-max_lines+1))
             for ((i=0; i<$max_lines; i++)); do
                 move_cursor "$((i+2));$((list_width+3))"
                 if [ $((yy+i)) -lt $listy_size ]; then
@@ -2484,14 +2491,16 @@ nsh() {
         }
         draw_shortcut_yank() {
             move_cursor $LINES
-            if [[ $op == 'mv' ]]; then
-                draw_shortcut "P" "Paste " "D" "Delete" "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
+            if [[ $op == cp ]]; then
+                draw_shortcut "p" "Paste " "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
+            elif [[ $op == mv ]]; then
+                draw_shortcut "p" "Paste " "D" "Delete" "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
             else
-                draw_shortcut "P" "Paste " "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
+                draw_shortcut "L" "Make Link" "/" "Search" "I" "Mkdir " "~" "Home  " "//" "Root  " "'" "Jump  "
             fi
         }
 
-        if [ ${#selected[@]} -eq 0 ]; then
+        if [[ ${#selected[@]} -eq 0 && $op != ln ]]; then
             if [[ ${list[$focus]} != ".." ]]; then
                 selected[$focus]="$PWD/${list[$focus]}"
             else
@@ -2636,31 +2645,42 @@ nsh() {
                     fi
                     draw_shortcut_yank
                     ;;
-                'p')
-                    if [[ $op == 'mv' ]]; then
-                        nshmv "${selected[@]/#$PWD\//}" "$path"
-                    else
-                        nshcp "${selected[@]/#$PWD\//}" "$path"
-                    fi
-                    #command cd "$path"
-                    last_item="$path/${list[$focus]}"
-                    update
-                    return
-                    ;;
-                'd')
-                    if [[ $op == 'mv' ]]; then
-                        if [[ ${#selected[@]} -eq 1 ]]; then
-                            dialog "Delete ${selected[@]/#$HOME/$tilde}?" " Yes " " No "
+                p)
+                    if [[ $op == cp || $op == mv ]]; then
+                        if [[ $op == 'mv' ]]; then
+                            nshmv "${selected[@]/#$PWD\//}" "$path"
                         else
-                            dialog "Delete ${#selected[@]} files?" " Yes " " No "
+                            nshcp "${selected[@]/#$PWD\//}" "$path"
                         fi
-                        [[ $? -eq 0 ]] && for f in "${selected[@]}"; do
-                            rm -rf "$f"
-                        done
+                        #command cd "$path"
+                        last_item="$path/${list[$focus]}"
+                        update
+                        return
                     fi
-                    update
-                    return
                     ;;
+                d)
+                    if [[ $op == cp || $op == mv ]]; then
+                        if [[ $op == 'mv' ]]; then
+                            if [[ ${#selected[@]} -eq 1 ]]; then
+                                dialog "Delete ${selected[@]/#$HOME/$tilde}?" " Yes " " No "
+                            else
+                                dialog "Delete ${#selected[@]} files?" " Yes " " No "
+                            fi
+                            [[ $? -eq 0 ]] && for f in "${selected[@]}"; do
+                                rm -rf "$f"
+                            done
+                        fi
+                        update
+                        return
+                    fi
+                    ;;
+                L)
+                    if [[ $op == ln ]]; then
+                        last_item="$PWD/${ylist[$yfocus]%/}"
+                        ln -s "$path/${ylist[$yfocus]}" "${last_item##*/}"
+                        update
+                        return
+                    fi
             esac
         done
         update
@@ -3404,39 +3424,23 @@ nsh() {
                 if [[ $STRING == *\ \& ]]; then
                     set -m
                 else
-                    set +m # ctrl-z doesn't work
+                    set +m # ctrl-z does not work
                 fi
-                eval "$STRING"$'\n'"$(echo NSH_EXIT_CODE=\$?)"
+                [[ -n $STRINGBUF ]] && STRING="$STRINGBUF"$'\n'"$STRING"
+                eval "$STRING 2>&1"$'\n'"$(echo NSH_EXIT_CODE=\$?)" 2>/dev/null
                 if [[ $? -ne 0 || -z "$__temp" ]]; then
-                    __string= ; print_prompt "$prefix"; hide_usage; echo
-                    __string="$__temp"
-                    [[ -n "$__string" ]] && echo -e "$NSH_PROMPT $(print_command $__string)"
-                    while true; do
-                        echo -en "$NSH_PROMPT "
-                        read_string --highlight ; echo
-                        if [[ -z "$STRING" ]]; then
-                            __manerr() {
-                                while read line; do
-                                    line="${line#$0: eval: line }"
-                                    echo "${line#* }"
-                                done
-                            }
-                            eval "$__string" 2> >(__manerr)
-                            NSH_EXIT_CODE=$?
-                            break
-                        fi
-                        __string="$__string"$'\n'"$STRING"
-                        tbeg=$(get_timestamp)
-                        eval "$__string 2>&1"$'\n'"$(echo NSH_EXIT_CODE=\$?)" 2>/dev/null
-                        if [[ $? -eq 0 ]]; then
-                            if [[ $secret -eq 0 ]]; then
-                                local tmp=$'\n'
-                                history[$((${#history[@]}-1))]="${__string//$tmp/ ; }"
-                            fi
-                            break
-                        fi
-                    done
+                    if [[ -z $STRINGBUF ]]; then
+                        move_cursor $row0 && echo -ne "$subprompt"
+                        hide_usage
+                        echo
+                        echo -ne "$NSH_PROMPT "
+                        syntax_highlight "$STRING"
+                        echo
+                    fi
+                    STRINGBUF="$STRING"
+                    return
                 fi
+                STRINGBUF=
                 local report=
                 if [[ $NSH_EXIT_CODE -ne 0 ]]; then
                     if [[ $STRING == git\ pull\ * && $STRING != *--force* ]]; then
@@ -3516,7 +3520,13 @@ nsh() {
                         select_all
                         ;;
                     $'\04')
-                        if [[ -z "$STRING" ]]; then
+                        if [[ -n $STRINGBUF ]]; then
+                            STRINGBUF=
+                            STRING=
+                            subprompt=
+                            echo '^C'
+                            break
+                        elif [[ -z "$STRING" ]]; then
                             STRING='exit'
                             NEXT_KEY=$'\n'
                         fi
@@ -3850,19 +3860,15 @@ nsh() {
                 echo "${bookmarks[$i]%% *}   ${val/#$HOME\//$tilde\/}"
             done; for ((i=$((${#visited[@]}-1)); i>=0; i--)); do
                 echo "    ${visited[$i]}"
-            done) | menu -h $max_lines --popup --header 'Key  Address' --footer $f --return-idx --searchable --key v 'echo edit')"
+            done) | menu -h $max_lines --popup --header 'Key  Address' --footer "$(draw_shortcut v Edit)" --searchable --key v 'echo \!edit')"
             if [[ $opened == yes ]]; then
                 disable_line_wrapping >&2
                 hide_cursor >&2
             fi
-            if [[ $i == edit ]]; then
+            if [[ $i == \!edit ]]; then
                 echo "$i"
             elif [[ -n "$i" ]]; then
-                if [ $i -lt ${#bookmarks[@]} ]; then
-                    echo "${bookmarks[$i]#* }"
-                else
-                    echo "${visited[$((${#visited[@]}-1-i+${#bookmarks[@]}))]/#$tilde\//$HOME/}"
-                fi
+                echo "${i#?}" | sed 's/^\ *//'
             fi
         else
             for ((i=0; i<${#bookmarks[@]}; i++)); do
@@ -4207,10 +4213,13 @@ nsh() {
                     fi
                     ;;
                 'y'|$'\e[15~') # cp
-                    yank
+                    yank cp
                     ;;
                 'd'|$'\e[17~')
-                    yank 'mv'
+                    yank mv
+                    ;;
+                L)
+                    yank ln
                     ;;
                 $'\e[12~'|$'\eOQ'|'i') # F2
                     f="${list[$focus]}"
@@ -4558,8 +4567,17 @@ nsh() {
                     ;;
                 "'"|'"')
                     local addr="$(select_bookmark "$KEY")"
-                    [[ "$addr" == edit ]] && subshell --secret "$NSH_DEFAULT_EDITOR ~/.config/nsh/bookmarks" && load_bookmarks && addr=
-                    [[ -n "$addr" ]] && open_file "$addr" || redraw
+                    if [[ "$addr" == \!edit ]]; then
+                        close_pane
+                        "$NSH_DEFAULT_EDITOR" ~/.config/nsh/bookmarks
+                        load_bookmarks
+                        open_pane
+                        update
+                    elif [[ -n "$addr" ]]; then
+                        open_file "$addr" || redraw
+                    else
+                        redraw
+                    fi
                     ;;
                 'q')
                     break
